+++ NEWS.md
@@ -1,4 +1,12 @@
 
+# rmonocypher 0.1.4  2024-05-03
+
+* Include functions for Shamir's Secret Sharing
+    * `create_keyshares(key, n, k)` to split a 32-bytes key into `n` pieces, 
+       and recoverable with only `k` of them.
+    * `combine_keyshares(shares)` to combine at least `k` shares in order 
+      to recover the original key.
+
 # rmonocypher 0.1.3  2024-04-26
 
 * Update `cryptfile()` to read/write with **connections** as well as files.